pyproject.toml
[project]
name = "agent-agentic-context-engineering"
version = "0.1.0"
description = "Simplified Agentic Context Engineering (ACE) using the Agent Framework"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agent-framework>=1.0.0rc2",
    "azure-identity>=1.19.0",
    "pydantic>=2.10,<2.13",
    "python-dotenv>=1.2.2",
]
